requirements.dev.txt
loguru==0.7.0
pillow==10.0.0
piexif==1.1.3
pytest==7.4.0
